+++ pkg/microservice/systemconfig/core/email/repository/mongo/emailservice.go
@@ -13,12 +13,6 @@ import (
 	mongotool "github.com/koderover/zadig/pkg/tool/mongo"
 )
 
-type EmailHostColl struct {
-	*mongo.Collection
-
-	coll string
-}
-
 type EmailServiceColl struct {
 	*mongo.Collection
 
@@ -32,82 +26,11 @@ func NewEmailServiceColl() *EmailServiceColl {
 	return coll
 }
 
-func NewEmailHostColl() *EmailHostColl {
-	name := models.EmailHost{}.TableName()
-	coll := &EmailHostColl{Collection: mongotool.Database(config.MongoDatabase()).Collection(name), coll: name}
-
-	return coll
-}
-
 func (c *EmailServiceColl) GetCollectionName() string {
 	return c.coll
 }
 func (c *EmailServiceColl) EnsureIndex(ctx context.Context) error {
 	return nil
-}
-
-func (c *EmailHostColl) GetCollectionName() string {
-	return c.coll
-}
-func (c *EmailHostColl) EnsureIndex(ctx context.Context) error {
-	return nil
-}
-
-func (c *EmailHostColl) Find() (*models.EmailHost, error) {
-	emailHost := new(models.EmailHost)
-	query := bson.M{"deleted_at": 0}
-
-	ctx := context.Background()
-
-	err := c.Collection.FindOne(ctx, query).Decode(emailHost)
-	if err != nil {
-		return nil, nil
-	}
-	emailHost.Password = "***"
-	return emailHost, nil
-}
-
-func (c *EmailHostColl) Update(emailHost *models.EmailHost) (*models.EmailHost, error) {
-	query := bson.M{"deleted_at": 0}
-	change := bson.M{"$set": bson.M{
-		"name":       emailHost.Name,
-		"port":       emailHost.Port,
-		"username":   emailHost.Username,
-		"password":   emailHost.Password,
-		"is_tls":     emailHost.IsTLS,
-		"updated_at": time.Now().Unix(),
-	}}
-
-	_, err := c.Collection.UpdateOne(context.TODO(), query, change)
-	if err != nil {
-		log.Error("repository Update EmailHostColl err : %v", err)
-		return nil, err
-	}
-	return emailHost, nil
-}
-
-func (c *EmailHostColl) Delete() error {
-	query := bson.M{"deleted_at": 0}
-	change := bson.M{"$set": bson.M{
-		"deleted_at": time.Now().Unix(),
-	}}
-
-	_, err := c.Collection.UpdateOne(context.TODO(), query, change)
-	if err != nil {
-		log.Error("repository Delete EmailHostColl err : %v", err)
-		return err
-	}
-	return nil
-}
-
-func (c *EmailHostColl) Add(emailHost *models.EmailHost) (*models.EmailHost, error) {
-
-	_, err := c.Collection.InsertOne(context.TODO(), emailHost)
-	if err != nil {
-		log.Error("repository AddEmailHost err : %v", err)
-		return nil, err
-	}
-	return emailHost, nil
 }
 
 func (c *EmailServiceColl) AddEmailService(iEmailService *models.EmailService) (*models.EmailService, error) {
